requirements.txt
openai>=1.0.0
pyaudio>=0.2.13
pynput>=1.7.6
pystray>=0.19.4
pillow>=10.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
pyautogui>=0.9.54
pytest==7.4.0
PyGObject 